requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
cycler==0.10.0
Cython==0.26
Flask==0.12.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
matplotlib==2.0.2
numpy==1.13.1
olefile==0.44
Pillow==4.2.1
pyparsing==2.2.0
pytesseract==0.1.7
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.2
six==1.10.0
urllib3==1.22
Werkzeug==0.12.2
